requirements.txt
allure-python-commons~=2.15.0
playwright~=1.55.0
pytest~=8.4.2
pytest-playwright~=0.7.1